added these refinements:
additions to document:

1. Message types are just typedef'd structures with no pointers
contained.  We won't allow subtyping of message types until later so
that we can simplify message dispatch for now.

2. The language syntax outside of our stream stuff is all of standard
C without pointers.

3. Everything passed via input and output of work functions is pass-by
value.

4. Re-initialization is done by saying "ref.reset(dynamic params...)"

changes to document:

1. There's no scripting "function" anymore, but rather just external
declaration of FilterBlocks.  Then, a FilterBlock can be instantiated
just like a Filter.  Then there's no ambiguity with functions
"returning" something; rather, all instantiations return a reference
that may or may not be bound to a variable name.

There can still be function declarations (in either the main stream
scope or in the scope of a single filter) that return values without
defining stream structures, just as in normal C.

2. The syntax of weighted round-robin changes to have no equals sign:

	WeightedRoundRobin = {1,2,3}	=>	WeightedRoundRobin(1,2,3)

3. A naming construct - you can give a string name to any instance
with the following syntax:

	FilterRef r1;

	r1 = MyFilter<1,2>(3) "this filter";

Then, the name can be used for identification or debugging by saying

	r1.getName()

The name returned is a hierarchical construction of all the filters
and filterblocks enclosing an instance.  If no name is given to an
instance at a given level, then a name is assigned according to the
type of the Filter/FilterBlock along with its position in each scope.

4. Clarification of semantics of stream control flow constructs: the
only constraint is that all enclosing constructs of statements which
construct filters or which modify scripting variables must be
s-<construct> type.

5. Split/join is extended to implicitly feed nothing into sources and
to expect nothing out of sinks.  Thus, if you have something like a
weighted split join where it specifies how many elements to send to
each block, the count must be zero for blocks that are a source.  This
implicit business allows us to introduce sources/sinks into the stream
in the middle, without having a split-join construct at the very
beginning or end.

6. Script parameters can be assigned to, so long as it's within script
control flow.  Also, you can have script control flow depend on script
functions, which are required to take only script parameters and to
return a script type, and not to reference anything dynamic within the
body of the function.

7. Input/output may appear in the initializer for setting up a
feedback loop, or getting the input tape rolling so that the Work
function can reference negative values.

left unchanged:

1. Let's change the name of Filters to Blocks and FilterBlocks to
Groups.  We think that Filters suggest something that's too
application-specific.

2. We think it'd be useful to leave in the "advance" construct in the
work function so that it's very clear to the programmer how much the
stream is being advanced with each invocation.  It's not a compiler
hint; it's just clear programming style.

------------------------------------------------------------------------

HAD THIS DISCUSSION:

stream ideas
------------

- do we still have flush/advance/push semantics for stream i/o?

- better idea of what's allowed within a message type

- let's present the function idea as a filterblock construct, not a
function.  then we can allow actual function calls in C that return
values and things... we can just define filterBlocks in a non-inlined
way as we could define filters in a non-inlined way.

- we need to specify the rest of the stream language that we allow.
all normal c syntax, pointers, etc?  what are the restrictions of
things that can be passed through the stream?  no pointers?  is it
that the body of the work function can't access non-local data (can't
point outside its address space?)

- I'd prefer a different syntax for weighted round robin.  something
without the equals sign.

- why is a split/join required before/after blocks without
output/input.

- so are we not worrying about automatic packing/unpacking anymore?
Might be a good idea, so that we don't have to automatically detect
the pack and unpack.

- are filters the right name for the base case, or should it be block
or something?

- I'd still like to add the naming stuff, so you can give an instance
of a block a name, and we can automatically trace the hierarchical
position of a block.

- need better idea of naming results of functions.  first say that
functions return nothing, but then we can take a reference of the
function's instantiation (but not the function itself, it says.)  Why
not just say that functions return a filterblock of the contents
inside, and that all filters and filterblocks return something when
they're instantiated, and then you can name this result if you want
to?  I don't like the "take a reference" concept.

- so pipe is basically a filterblock w/o variables & scoping

- I don't understand the last example in section 4.6 -- what's up with
the split/join syntax and stuff?

- i think sif/sfor should change so that sif/sfor is used whenever
blocks are created inside the constructs.  then you can use a normal
if for script variables if you don't define things inside, right?


little things
-------------

- what does it mean "a script type can only be assigned a constant
value, and can only be modified within script control constructs"?


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@48 fb122d4e-f3b4-4596-99d2-990b948fa53d