I think the Flattener is completely done (!), except for one detail with the buffer size interface that I can easily work out with Michal.
So try "java at.dms.kjc.sir.SIRBuilder" to see (what I think is) the
low-ir output of HelloWorld6.

I added an LIRSetTape construct to specify the tapes to the runtime
system.  I'm assuming right now that tapes are stored in the
structures/closures input to functions so that I don't have them
around very much.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@351 fb122d4e-f3b4-4596-99d2-990b948fa53d