The constant propagator works now.  I think the only thing it's lacking is constant propagation through fields--unfortunatly, if an anonymous class references a final variable in an enclosing scope, then it does this as a field reference instead of a local var reference.
So we'll probably want to add this, although I'll move on to loop
unrolling for now... it should be possible to write FFT without the
references to enclosing final vars.

Also, we'll need to adjust where round-robin weights are specified.
Right now it's in Kopi2SIR, but it'll have to come after constant
propagation, hmmm... will need some decoupling.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@619 fb122d4e-f3b4-4596-99d2-990b948fa53d