This version of FFT now runs from end to end!!  The bad news is that it gets slightly the wrong answer.  It has the periodicity right, but it looks like it's off by a few (compared to the library, at least--might be worth checking them both.)
I had to remove the array of weights, since we don't deal with arrays
yet.  Otherwise the only difference between this and FFT.java is some
extra names for classes, declcaring variables at the beginning of
a procedure, and changing ++ to +=1.

So this is very cool... it's a thousand lines of C code in all that is
running!  wohoo!

I'm going to time it now and see how quickly we get
almost-the-right-answer.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@673 fb122d4e-f3b4-4596-99d2-990b948fa53d