Adding a C version of the GSM vocoder that Jeremy found.  I'm sorry for all the add messages--I don't know any better way to add files when things are mixed between binary and text.  If you want to add a whole directory structure that's either all binary or all text, you can use import, like I did for the test cases.
I always get the import syntax wrong.  Here's a clear example from a
CVS manual:

-----------------------------------------------------------------------

When you begin using CVS, you will probably already have several
projects that can be put under CVS control. In these cases the easiest
way is to use the import command. An example is probably the easiest
way to explain how to use it. If the files you want to install in CVS
reside in `wdir', and you want them to appear in the repository as
`$CVSROOT/yoyodyne/rdir', you can do this:

$ cd wdir
$ cvs import -m "Imported sources" yoyodyne/rdir yoyo start

Unless you supply a log message with the `-m' flag, CVS starts an
editor and prompts for a message. The string `yoyo' is a vendor tag,
and `start' is a release tag. They may fill no purpose in this
context, but since CVS requires them they must be present. See section
13. Tracking third-party sources, for more information about them.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@741 fb122d4e-f3b4-4596-99d2-990b948fa53d