Here's the current version of the CC paper.  It is MAJORLY revised over what Saman sent earlier this week... most notably, it has the bit about the common machine language, a re-written intro, conclusion, and filter composition section, new splitjoin example, lots of formatting, and some other stuff.
Supposedly this was due on Friday, but I told the Chair that I'd send it
to him this morning, and I haven't heard back from him yet.  There's still
a lot to do -- have to get it down to 17 pages, need to rewrite the detailed
example [I just truncated it to the length I hope it'll end up being] and
take care of the other stuff in the "todo" file.  So I'll keep scrambling
'til I hear from him, which hopefully won't be too long.

Please let me know if you catch any bugs or have good suggestions on what
to cut out!  I want to make room to add an acknowledgements section with
everyone's name.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1125 fb122d4e-f3b4-4596-99d2-990b948fa53d