Makefile includes new files.
SIRFilter has a toString method, plus ways to set peek/pop/push as
ints instead of only JExpressions.

Removed debug output from SIRPeekExpression.

SIRPipeline has less debug output.

SIRTwoStageFilter now has the correct signature on its copyState
function, so that its implementation overrides the one in SIRFilter.
Note that if you extend a class and want to override a method, that
method needs to have not only the same name and return type, but it
can NOT have a MORE specific argument type in the subclass if you want
the method to be dynamically dispatched to.  Thanks to David for
catching this.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1180 fb122d4e-f3b4-4596-99d2-990b948fa53d