AdjustGranularity has my latest granularity stuff.
Flattener includes a call to the Partitioner if (-a) is specified.

WorkEstimate has methods to return a sorted list of filters by amount
of work, as well as a list of containers that is sorted by the amount
of work of immediate children.

Partitioner is the automatic partitioner.

WorkInfo is just an integer wrapper (at this point) for describing
work.  WorkList is an extension of LinkedList for easy interfacing
with work amounts.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1449 fb122d4e-f3b4-4596-99d2-990b948fa53d