Fixing two bugs in our FFT implementation, that were emailed to us from a master's student at Berkeley.  This is kind of embarassing... we definately need some reference implementations ASAP.
I'll correct the CC paper when I get around to it, but unfortunately
the bug also causes a change in the structure of the stream graph,
which will affect the ASPLOS numbers.  So we should re-do them when we
have the chance or someone feels inspired (gordo?)

Also, I changed FFT to get the same answer as FFTLocal.  There were
some weird differences between the two that were making their answers
different before.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1496 fb122d4e-f3b4-4596-99d2-990b948fa53d