Reed-Solomon encoding/decoding example. Works with the library.
The implementation of the heavy-lifting routines was lifted from the internet (see comments).

The Reed-Solomon code is really slick -- not only is it used in the HDTV decoder, it is also used in things like CDs and cell phones. It is like the simple ECC that they teach in 6.004 on crack.

With the addition of k parity bytes, upto k/2 errors can be successfully found and corrected. If you know the locations of the errors (erasures) you can correct k errors.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1575 fb122d4e-f3b4-4596-99d2-990b948fa53d