This commit makes BitonicSort get the right answer with the compiler.
The most EVIL fix is to JLocalVariable, where the KOPI folks had
overriden the hashCode() function to just hash the name and index of
the variable.  This was causing ConstProp to propagate by name, not
object identity.  Changed back to default Object.hashCode().

Added .equals methods for all the literal types.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1664 fb122d4e-f3b4-4596-99d2-990b948fa53d