IR tree constructs that were being recreated from the original parse tree (eg with new JVariableDeclaration(...) were not passing through comments and token references.
I went though and added self.getTokenReference() and self.getComments(), replacing null in the constructors appropriately.

I am hoping to use special comments in the near future for a hack-ish hook into the compiler to let me declare variables as "bit"s, and not have to deal with grammar hacking or screwing with java compatability.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1798 fb122d4e-f3b4-4596-99d2-990b948fa53d