Substantial overhaul of the raw backend.  Most importantly, I am confident that we now have a correct, deadlock-free implementation of feedbackloops. It will not deadlock as long as the data returned from the initPath() calls will fit in tile memory.
There were also some graph arrangements that would lead to a crash.  For
example, an application that was just a feedbackloop would fail.
Having a joiner with a null incoming arc was a problem...
I think that now ALL legal streamit graphs will work.

Also, I went ahead and comment and rewrote some of the ASPLOS deadline code
that was impossible to understand.  There is still lots more of it though.

I tested the changes, but we will have to wait for the next regtest email
to see if it is solid.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1836 fb122d4e-f3b4-4596-99d2-990b948fa53d