Two Changes:
To fix the problem with var names clashing between params to the init function
and fields (ex. Lattice), the RawBackend now runs the renamer.  Let's wait
for the next regtest to see if this breaks anything.

Also, when arrays are passed to init functions, we constant prop the
elements of assignments, so the arrays are not needed.  So, now we
no longer pass try to pass the array.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@1837 fb122d4e-f3b4-4596-99d2-990b948fa53d