This commit adds the following functionality to SplitJoin fusion:   * Fusion of two-stage filters.   * Fusion of filters that peek, w/ any kind of round-robin splitter.
If this change causes any problems, you can revert to only allowing
the old kind of fusion with the -S flag to the compiler.  The old
fusion is simpler and somewhat more efficient, so the new fusion
already employs it whenever it can.  But -S will disable the new
fusion entirely.

I tested this on everything I could find and it seems to be correct
now.  It fuses the 128-Vocoder down to 16 tiles (although the
Simulator gives an error when trying to run.)  Should also fix Nokia's
fusion woes.

---

More detailed comments:

- JBlock has a copy() method.

- Options files includes -S for reverting to Simple SplitJoin fusion.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@2145 fb122d4e-f3b4-4596-99d2-990b948fa53d