Big change in how the raw backend works.  I love a big change right before the deadline.  This has not been tested too well, but it handles everything that broke the old ver.  If this breaks anything, you will have to revert to the old ver.  This check in will not change execution time, so numbers do not have to be rerun.
I will not go into the change here, just know that the old ver will
deadlock in certain cases.

Also, the print code that dave wrote for me did not work for 64 tiles.  I
reverted to the older code.  It is slow, use a file writer if there is
a lot of printing in the steady state.

NEVER USE -x it is broken.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@2226 fb122d4e-f3b4-4596-99d2-990b948fa53d