LinearDot: now colors nodes differently based on if they have constant components or not.
AccessWrapperFactory and LinearFilterAnalyzer: modifications to support arrays and array initializations. Eventually some of this will be yanked out because ConstProp will handle it, but until then it is needed for some of the programs.

LinearFilterDirectReplacer: Code for replacing the work function of streams that we have linear representations for with a direct implementation (untested currently).


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@2553 fb122d4e-f3b4-4596-99d2-990b948fa53d