*** IMPORTANT PART OF CHECK-IN ***
ConstructSIRTree now REMOVES init-function arguments that are constant
after the const. prop / graph expansion phase.  At this point, this
removes all arguments to init, since we assert that arguments are
constant.

This fixes parts of the regtest (at least FFT3) where a fused init
function was referencing an argument that no longer existed due to
either constant prop, or partitioning (e.g. if the declaration was on
a different RAW tile.)

----------------------------------

RenameAll now finds new variable declarations from the initialization
part of FOR loops.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@2628 fb122d4e-f3b4-4596-99d2-990b948fa53d