Made a first attempt to clean this up.  Only found one actual parser error (not supporting unary minus).  Fixed a lot of other things that might have been syntactically correct Java but weren't anywhere near legal new syntax.
This is still pretty broken.  I think CubVect (line 47) is a structure
type, maybe?  But there's no definition.  Similarly, LQ appears to be
intended to be a helper class, but (a) we have no support for helper
functions at all and (b) the code is missing in any case.  It looks
like ../FAT/LQ.java exists, so if helper function support happens
there's at least some code.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@2821 fb122d4e-f3b4-4596-99d2-990b948fa53d