I added a pass to convert global variables into local variables to (hopefully) help gcc better schedule the generated code for the superscalar raw simulator...
It inlines all function calls that it can, so that it ends up with one large
main function that includes the init(), preWork(), and work() and anything
they call inlined in the code.  All the globals become locals of this main
function.  If it cannot inline a function call (the call is embedded in a
compilicated expression) then if will still perform the conversion if the
caller does not access globals.  It will not perform the conversion if there
are mutually recursive function calls (this is almost the case now, I have
to finish this part, right now it is overly cautious).  There is some
other crap but that should be all you need to know.

To run it, use -G (or --removeglobals).  The pass is a bit slow
 and I do not know why, I do use a lot of visitors, but they should
not be this slow.  I will look at it.  It works for the apps the I have
thrown at it (a bunch of  examples, Beamformer, and FM).  I will keep
testing it.

I do not know if this should be run by default. I will have to check the
performance versus the default code.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3031 fb122d4e-f3b4-4596-99d2-990b948fa53d