Looks like the changes to raw codegen are ready for mainstream. It does away with inline volatile instructions that were preventing gcc from optimizing and scheduling freely. Instead it does a hack involving:
global registers router and router_f tied to $24
global router_mem and router_mem_f to act as guards to protect from optimizations
inline assembly (nonvolatile) that uses router/router_f (that also are specified to clobber $24 so $24 isn't reused)
hacking at the Makefile to remove references to router_mem/router_mem_f from tile*.s before proceeding with building

Nothing less (and doable within time) seems to get the job done. Some gcc optimization or another clobbered everything else.
This guy better help out:>

I tested it with all the benchmark apps and it seems I finally settled on a combination that works..
The last to fall was LinkedFM which took way too long and was just due to a lack of casting.
However, it probably isn't quite ready to be default.
In the tradition of Bill, I gave it the option -z or --sketchycodegen
The option is only suppose to be temporary. Hopefully it will eventually be stable enough to become default in some form or another.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3073 fb122d4e-f3b4-4596-99d2-990b948fa53d