Wrote the paragraph on helper functions.  Syntax is return-type [const] helper name (params) { body }; helpers explicitly declared const may be optimized away, but can't change filter state or call non-const helpers.
Make the semicolon after anonymous stream declarations optional; both

  add int->void filter { work pop 1 { pop(); } };

and

  add float->float filter {
    ...
  }

are legal.

Eliminate the example syntax for phased filters, since this is a
little controversial and we've seen better syntaxes on the mailing
list.

Add a short section on variable rates as something that's not
supported.

Add a section on implementation limits, which right now has the pop
confusion and a list of names that are reserved due to being used in
the Java form or the C back-end.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3199 fb122d4e-f3b4-4596-99d2-990b948fa53d