New script that runs the regtest each evening. Note that it is a perl script this time.
The behavior is as follows:
1. Check out clean streams source tree in /home/bits6/NO_BACKUP/streamit/regtest_working/datestamp
2. Do our crazy build process (eg make;clean;compile;make;clean;compile).
3. Run the regtest framework (specifically the tests in TestAll.java -- I am thinking of renaming this so that it is something like "RegTest"...)

One thing that I didn't get working (because I was lazy) was the "what apps aren't in the regtest" thing that Bill had going. I need to fix that, but I am not sure what it takes as input, so I am waiting for tomorrow.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3287 fb122d4e-f3b4-4596-99d2-990b948fa53d