run_reg_test should now be good to go -- executed each evening at 10PM on cagfarm-47. Bill has suggested that we have the regtest run every 12 hours, which is feasible now that each regtest instance is run in its own directory.
One thing to note about the timeouts: Currently they will be reported as failures without any special annotations that they timed out. I will have all the output emailed to me so I can figure out which ones are timing out manually. If a good many of the tests are timing out, I will go and add something that marks the result of a specific test as having failed because it timed out.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3296 fb122d4e-f3b4-4596-99d2-990b948fa53d