Adjusting the partitioner record-keeping to keep track of work in fissed nodes.
PartitionRecord is a new bookkeping unit for a partition, keeping
track of work and contents of a partition.  Changing PartitionUtil to
print straight to file when saving statistics; changing
DynamicProgPartitioner to use PartitionRecord's instead of ad-hoc
hashtables and such.

PartitionDot also outputs list of tiles assigned to a fissed filter.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3364 fb122d4e-f3b4-4596-99d2-990b948fa53d