The Jan. 9 #1 version sent to Peter.
my own changes
--------------

- fixing inheritance rules -- making inheritance from base_stream
virtual so that base object shared in multiple inheritance hierarchy;
adding "= 0" to indicate pure virtual function on work, work_info.

- fixing mergesort work function to deal with EOS correctly

- changing draining semantics so that it doesn't fire unless peek_rate
items are available.  This is a common case with a static peek rate.
If there is really a variable peek rate, it should be declared with
dynamicPeek.  (then as we discussed, draining is until it returns true
or fewer than peek items left on stream.)

- removing default value of peek to be pop, since it's mildly unclear
given that pop can be static or dynamic

changes as discussed w/ Peter, with my own modifications
--------------------------------------------------------

- reversing the sign on the boolean returned from work or prework.
It's easier to describe if it returns "whether or not its finished"
instead of "whether or not it should be fired again", since the latter
also depends on the number of input items in the stream.  I hope it
doesn't look too unusual to have a bunch of "return false" statements
floating around.

- changed PROC_MEMBANK to PROC_DMEMBANK, then added PROC_IMEMBANK and
PROC_ISIZE.

- adding protected access modifiers in addition to private, public

changes as discussed changes w/ Peter
-------------------------------------

- all arguments to stream constructor are literals.  all processor
resources are always literals.

- removing SRF mem node type, should count as "regular" data memory

- kernel state and length of streams undefined during execution


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3413 fb122d4e-f3b4-4596-99d2-990b948fa53d