Draft sent to Peter on Jan. 9, #2.  Changes:
new, not discussed
------------------

- hairy corner case: a graph has a stream as an output, but it so
happens that no items are written to that stream during execution.
Then is the length reset to zero?  I assumed yes.

- added a footnote about removing two dynamic restrictions to allow
morphable / resumable graphs.

as discussed with caveats
-------------------------

- the ostream.of(int index) function only returns true if the stream
is not circular.  Then a kernel stops not when all its output streams
are "full", but either when a blocked static-rate output stream is
full, or when all of its downstream kernels have terminated.  (Or when
it returns "true" from work.)

  -> otherwise would have scheduler-dependent behavior sometimes

as discussed
------------

- adding memory locations to examples

- changing "write_once" to "circular" as argument to stream constructor

- adding of(int index) function to ostream

- circular streams don't have overlapping memory segments

- any ordering of kernels according to data dependences is fair game for the low-level compiler

- HLC has to insert waits if there are memory-location dependences between graphs


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3431 fb122d4e-f3b4-4596-99d2-990b948fa53d