Moving CUPLexParseactions into its own class, and removing the $ in its name, so that the build process will work with the 3rdparty directory.
Call me crazy, but it looks like jikes +M gives erroneous output for
inner classes when it is used from a directory other than the base of
a package hierarchy.  That is, there are different .u files generated
for these commands:

jikes +M 3rdparty/JFlex/Main.java
cd 3rdparty && jikes +M JFlex/Main.java

In the former case, Main.java is said to depend on a SOURCE file of
the inner class of LexParse.java, which it doesn't.

So, for now I'm making this change to the JFlex sources so our build
process will work.  If this offends anyone then later we can figure
out what's wrong with jikes (or my impression of it).


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3459 fb122d4e-f3b4-4596-99d2-990b948fa53d