The java code that implements magic network support in its current form.
First off, the magic network does not work for everything.  bC is just so
quirky (buggy?).  I do not think it was meant to head the large files that
I am generating to represent the communication schedule.  Mallocs are
failing at random times and other weird behavior.  One more thing I will
say, if you are ever going to write some bC, do not place any statements
in the global scope (other than function calls).  Sequences of code that
die in the global scope run fine embedded in a function.

A quick summary of the implementation.  These files generate bC data
structures that represent the receive schedule and the send schedule for
each tile.  The schedules is represented as a linked list with the last
node pointing to the start of the steady state.  So, when the simulator
starts up, these structures are initialized, and as the simulation runs
each linked list is traversed for the schedule of sending or receiving.
There are a lot of implementation details but that should be enough.

I hope after some more fiddling I can get it stable...


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@3555 fb122d4e-f3b4-4596-99d2-990b948fa53d