Now the number gathering infrastructure will work with most benchmarks that have multiple printers and sinks.  But there are some limitations. The multiple sinks must be "synchronized".  By this I mean:
In streamit the set of sink nodes has a single least common ancestor.  I
check that all filters from each sink to the least common ancestor
have a pop rate > 0 (they are synchronized to the previous filter).

I also count the prints from all the sinks in the thruput calculations.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@4222 fb122d4e-f3b4-4596-99d2-990b948fa53d