Fixing rt ticket 103.  -G (remove globals) broke GSM.
The problem was with a function call that was an arguments to a
function call.  One of the visitors I wrote was not visiting the
arguments of function calls, so the remove global pass had no
knowledge of the function call in the argument.  So it was not
generating the function declaration (it could not inline them).  So
after wasting a bunch of time, it is fixed.  One line.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@4309 fb122d4e-f3b4-4596-99d2-990b948fa53d