I am finally implementing a correct version of rate matching.
With this update the check to see if we can perform rate matching is
implemented.  Basically it checks to see if there are any overlapping
routes or routes that go thru tiles assigned to nodes.

Now I have to implement the pass which will require changes to the
simulator and the raw code generation passes.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@4568 fb122d4e-f3b4-4596-99d2-990b948fa53d