BUG Fix.  The remove globals pass does not mesh too well with multidimensional arrays.  It is so hard converting java multidim arrays to C.  There is once case that led to some crazy behavior that was hard to track down.
Defining a multidimensional array in StreamIt, then not declaring it.
Instead just assigning to it a declared array.  In C, if you assign a
stack allocated multidim array to a ** (pointer to a pointer) the **
is pointing to garbage.

So the RemoveGlobals pass checks for any multidimensional arrays that are
not allocated.  If there exists one, we do not perform the conversion.

I could actually get this to work, but it would require a lot of
reorganization and special casing.  It only shows up in one filter
in Nokia, so I am punting.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@4661 fb122d4e-f3b4-4596-99d2-990b948fa53d