Bug fix:
This bug took me a while to find and I am surprised that it did not
come up sooner.  It might be a by-product of the new scheduler, I do
not know.

In the initialization phase, the number of items produced by a
splitter may not equal the number of item produced by the filter
feeding the splitter.  The items are in the input tape of the
splitter.

Now, since I do not map splitter, there is no input buffer for the
splitter.  This discrepancy must be accounted for.  We do not have an
incoming buffer for ths splitter, so the data must trickle down to
the filter(s) that the splitter feeds.

The simulator does the right thing, because it now ignores the number
of times that the splitter should fire in the initialization stage.
But the tile code of the downstream filters of the splitter must know
they are to receive this extra data that is not in the schedule.

So I added a bunch of code that does this.  It recognizes if there
is code stored up in a splitter's input buffer in the initialization
stage, and if so, it lets the downstream filter(s) know that they
should receive this extra data.  Also, there might be many splitters
that need to be passed thru before the data gets to the filter(s) and
we must calculate the exact number of items produced by the filter
feeding the top splitter and passed to each downstream filter.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@4662 fb122d4e-f3b4-4596-99d2-990b948fa53d