Jasper figured out a 4-line improvement to the Beamformer code that improves performance by 25%.  It reverses the order of loop indexing in the internal array, thereby allowing us to use an and expression for wrapping around instead of using a branch.
We could call this beamformer2 if we want to distinguish it from our
previous benchmark... but I think it might be okay to edit the source
in this case, since our previous version was just naive, and this is
medium-intelligent.  Moreover, we have to compare the whole
application with lincoln labs, etc. where theirs is written in
hand-optimized C.  Let's talk about this if anyone has objections.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@4669 fb122d4e-f3b4-4596-99d2-990b948fa53d