This is a stable version of the 'cycle accurate' communication scheduler. When scheduling communication on Raw's static network, this scheduler will take into account the work estimation of the tiles and set the firing order accordingly.  So this enables tiles to route while they are computing. Before a tile could only route thru it in between firings.  This helps a lot on certain applications.  It should not deadlock ever...
For the curious.  It is implemented as an event-based simulation, where
there is only one type of event, the sending of one item.  That is
all I feel like saying about the implementation, it is actually well
documented.

Oh, it is not enabled by default yet, --wbs will enable it. wbs stands
for work-based simulator.  Don't ask.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@4919 fb122d4e-f3b4-4596-99d2-990b948fa53d