Removing a few semantic checks from Kopi, because they were broken and reject legal Java.
In the translated version of BeamFormer.str, JLocalVariableExpression
was complaining about a possibly uninitialized local variable, and
JAssignmentExpression was disallowing final variable assignments in
loops [they're legal if the variable is also declared in the loop.]

Jikes accepts both these cases, so I'm declaring Kopi as broken and
commenting out the checks so our benchmarks go through.  We should
detect such things in the frontend.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@5053 fb122d4e-f3b4-4596-99d2-990b948fa53d