This checkin fixes ticket 169.  The problem was that Layout is a static class (all the state is static) so the wbs called layout late in the compiler flow and it quashed the layout of the app.
Now what I should have done was make layout store its state in
instance vars.  But that would take forever and I do not care too
much about the old backend.  So, now there are methods to save/get the
state of layout thru an internal layout field.  All the state
in layout is now an instance of this field.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@5114 fb122d4e-f3b4-4596-99d2-990b948fa53d