Three fixes the should reduce regtest errors.  First I was not catching all the math function calls and converting them to single precision (somehow sin got by me) but I made and list and checked it twice.
Next, my array-over-tapes implementation was not jiving with the
array destroyer because a it was not in the IR.  I now place
all array receive code in the IR and let the array destroyer
do with it as it may.

Finally, we need to dynamically allocate local arrays in the init
function to support init arguments propagation.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@5215 fb122d4e-f3b4-4596-99d2-990b948fa53d