So, I finally found the problem with FMRadio!  It was a problem with the application, not the compiler (well, maybe, read on).  It was a problem with the input filter:
void->float filter FloatOneSource {

  int x;

  init {
    x = 0;
  }

  work push 1 pop 0 {
    push(x++);
  }
}

Anyone see the problem?  Come on.

The front-end should really check that the type of push statement is the
same type as the filter output type.  I added a check to the raw backend
and I will check it in.  I will also add a check to Kopi2SIR if I can.

This was a problem in FloatOneSource also.


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@5273 fb122d4e-f3b4-4596-99d2-990b948fa53d