Some more weirdness.  I ran Vocoder thru after I put the check into kopitosir (for the type of the argument for a push statement) and it failed.  The filter that failed is:
class IntToFloat extends Filter {
  public void init() {
    input = new Channel(Integer.TYPE, 1);
    output = new Channel(Float.TYPE, 1);
  }
  public void work() {
    output.pushFloat((float)input.popInt());
  }
}

It looks fine with the cast in there.  But when kopitosir sees it, the
cast is missing.  Hmmm.... This has to be fixed!

So, I removed the check from kopitosir and in FlatIRToC I add an explicit
cast if the type of the arg to the push does not match the filter output
type.

Now, the answer for Vocoder on raw matches the uniprocessor, but they both
differ from the library.  Weird.

Can someone look at Kopi to see why it would remove the cast above.  What
should we do about raw/uni differing from the lib?


git-svn-id: https://svn.csail.mit.edu/streamit/trunk@5278 fb122d4e-f3b4-4596-99d2-990b948fa53d